Fix: rename extension 'toast' to 'banner'
