+++ apps/web-client/src/lib/components/ExtensionStatusBanner.svelte
